[#2145] Fixed installer script requiring `curl`.
